PIP 모드 추가, 모놀리식 서버로 교체, 동영상 가로/세로 구분

+++ design-system/src/main/java/com/danbam/design_system/component/TabLayout.kt
@@ -206,7 +206,7 @@ fun IndiStrawRowTab(
     itemList: List<MovieEntity>,
     tabHeader: @Composable () -> Unit,
     moreData: (() -> Unit)? = null,
-    onClickItem: (Int) -> Unit,
+    onClickItem: (Long) -> Unit,
 ) {
     val state = rememberLazyListState()
