+++ domain/src/main/java/com/danbam/domain/entity/MovieEntity.kt
@@ -1,6 +1,6 @@
 package com.danbam.domain.entity
 
 data class MovieEntity(
-    val idx: Int,
+    val idx: Long,
     val thumbnailUrl: String,
 )
