+++ domain/src/main/java/com/danbam/domain/entity/MovieHistoryEntity.kt
@@ -1,7 +1,7 @@
 package com.danbam.domain.entity
 
 data class MovieHistoryEntity(
-    val movieIdx: Int,
+    val movieIdx: Long,
     val thumbnailUrl: String,
     val historyTime: Float,
     val title: String,
