+++ domain/src/main/java/com/danbam/domain/usecase/movie/MovieHistoryUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 class MovieHistoryUseCase @Inject constructor(
     private val movieRepository: MovieRepository
 ) {
-    suspend operator fun invoke(movieIdx: Int) = runCatching {
+    suspend operator fun invoke(movieIdx: Long) = runCatching {
         movieRepository.movieHistory(movieIdx = movieIdx)
     }
 }
