+++ tv/src/main/java/com/danbam/tv/ui/home/HomeContract.kt
@@ -4,7 +4,7 @@ import com.danbam.domain.entity.BannerEntity
 import com.danbam.domain.entity.MovieEntity
 
 data class HomeState(
-    val currentMovieIndex: Int = 0,
+    val currentMovieIndex: Long = 0,
     val movieList: List<MovieEntity> = listOf(),
     val bannerList: List<BannerEntity> = listOf()
 )
